relocate templates dir, remove url prefixing, add blueprint
