+++ cpm_app/__init__.py
@@ -4,19 +4,15 @@ from flask import Flask, render_template
 
 
 def create_app():
-    app = Flask("CPM", instance_relative_config=True)
+    app = Flask(__name__, instance_relative_config=True)
 
     # maximum file size 25 MB
     app.config.from_mapping(
         SECRET_KEY=os.getenv("SECRET_KEY"), MAX_CONTENT_LENGTH=25 * 1024 * 1024
     )
 
-    if __name__ == "__main__":
-        app.run()
-
-    from . import views
-
-    views.register_routes(app)
+    from . import upload
+    app.register_blueprint(upload.bp)
     app.add_url_rule(rule="/", endpoint="index")
 
     return app
