+++ cpm_app/templates/base.html
@@ -7,11 +7,12 @@
     <link href="css/style.css" rel="stylesheet" />
   </head>
   <body>
+    <header>{% block header %}{% endblock %}<header />
     <nav>
-      <h1>CPM</h1>
       <a>About</a>
       <a>Github</a>
     </nav>
+    <br/>
     <section>
       <span>Upload your image file to generate a color palette!</span>
       <div>
