+++ cpm_app/templates/index.html
@@ -1,0 +1,4 @@
+{% extends 'base.html' %}
+
+{% block title %}home{% endblock %}
+<h1>{% block header %}CPM{% endblock %}<h1/>
