--- templates/index.html
@@ -1,5 +1,0 @@
-{% extends 'base.html' %}
-
-{% block header %}
-  <h1>{% block title %}Color Palette Maker{% endblock %}</h1>
-{% endblock %}
